[IMP] stock_release_channel_plan: tests
Extract setup in common for inheritance
